+++ security-jwt/src/main/java/io/micronaut/security/token/jwt/cookie/JwtCookieConfigurationProperties.java
@@ -17,10 +17,14 @@ package io.micronaut.security.token.jwt.cookie;
 
 import io.micronaut.context.annotation.ConfigurationProperties;
 import io.micronaut.context.annotation.Requires;
+import io.micronaut.core.annotation.NonNull;
+import io.micronaut.core.annotation.Nullable;
 import io.micronaut.core.util.StringUtils;
 import io.micronaut.http.cookie.SameSite;
 import io.micronaut.security.authentication.CookieBasedAuthenticationModeCondition;
 import io.micronaut.security.token.jwt.config.JwtConfigurationProperties;
+
+import java.util.Optional;
 
 /**
  *
@@ -30,7 +34,7 @@ import io.micronaut.security.token.jwt.config.JwtConfigurationProperties;
 @Requires(condition = CookieBasedAuthenticationModeCondition.class)
 @Requires(property = JwtCookieConfigurationProperties.PREFIX + ".enabled", notEquals = StringUtils.FALSE, defaultValue = StringUtils.TRUE)
 @ConfigurationProperties(JwtCookieConfigurationProperties.PREFIX)
-public class JwtCookieConfigurationProperties extends AbstractAccessTokenCookieConfigurationProperties {
+public class JwtCookieConfigurationProperties extends AbstractAccessTokenCookieConfigurationProperties implements AccessTokenCookieConfiguration {
 
     public static final String PREFIX = JwtConfigurationProperties.PREFIX + ".cookie";
 
@@ -56,11 +60,9 @@ public class JwtCookieConfigurationProperties extends AbstractAccessTokenCookieC
 
     /**
      * Default Cookie Path.
-     *
-     * @deprecated use {@link AbstractAccessTokenCookieConfigurationProperties#DEFAULT_COOKIEPATH}
      */
-    @Deprecated
-    public static final String DEFAULT_COOKIEPATH = AbstractAccessTokenCookieConfigurationProperties.DEFAULT_COOKIEPATH;
+    @SuppressWarnings("WeakerAccess")
+    public static final String DEFAULT_COOKIEPATH = "/";
 
     /**
      * The default same-site setting for the JWT cookie.
@@ -70,9 +72,18 @@ public class JwtCookieConfigurationProperties extends AbstractAccessTokenCookieC
     @Deprecated
     public static final SameSite DEFAULT_COOKIESAMESITE = AbstractAccessTokenCookieConfigurationProperties.DEFAULT_COOKIESAMESITE;
 
-    protected boolean enabled = DEFAULT_ENABLED;
-    protected String cookieName = DEFAULT_COOKIENAME;
+    private String cookiePath = DEFAULT_COOKIEPATH;
+    private boolean enabled = DEFAULT_ENABLED;
+    private String cookieName = DEFAULT_COOKIENAME;
 
+    /**
+     *
+     * @return a boolean flag indicating whether the JwtCookieTokenReader should be enabled or not
+     */
+    @Override
+    public boolean isEnabled() {
+        return enabled;
+    }
 
     /**
      * Sets whether JWT cookie configuration is enabled. Default value ({@value #DEFAULT_ENABLED}).
@@ -90,4 +101,31 @@ public class JwtCookieConfigurationProperties extends AbstractAccessTokenCookieC
         this.cookieName = cookieName;
     }
 
+    /**
+     *
+     * @return a name for the cookie
+     */
+    @NonNull
+    @Override
+    public String getCookieName() {
+        return this.cookieName;
+    }
+
+    /**
+     *
+     * @return The path of the cookie.
+     */
+    @Nullable
+    @Override
+    public Optional<String> getCookiePath() {
+        return Optional.ofNullable(cookiePath);
+    }
+
+    /**
+     * Sets the path of the cookie. Default value ({@value #DEFAULT_COOKIEPATH}).
+     * @param cookiePath The path of the cookie.
+     */
+    public void setCookiePath(@Nullable String cookiePath) {
+        this.cookiePath = cookiePath;
+    }
 }
